@@ -21,155 +21,153 @@
     "url": "https://github.com/webpack/webpack.js.org/issues"
   },
   "engines": {
-    "node": ">=8.9.4"
+    "node": ">=10.13.0"
   },
   "scripts": {
     "clean-dist": "rimraf ./dist",
     "clean-printable": "rimraf src/content/**/printable.md",
     "preclean": "run-s clean-dist clean-printable",
-    "clean": "rimraf src/content/**/_*.md src/**/_*.json",
-    "start": "npm run clean-dist && cross-env NODE_ENV=development webpack-dev-server --config webpack.dev.js --env.dev",
-    "update-repos": "node src/utilities/fetch-package-repos.js",
+    "clean": "rimraf src/content/**/_*.md src/**/_*.json repositories/*.json",
+    "start": "npm run clean-dist && webpack serve --config webpack.dev.js --env dev --progress --node-env development",
     "content": "node src/scripts/build-content-tree.js ./src/content ./src/_content.json",
-    "bundle-analyze": "run-s clean fetch printable content && cross-env NODE_ENV=production webpack --config webpack.ssg.js && run-s clean-printable content && cross-env NODE_ENV=production webpack --config webpack.prod.js  --profile --json > stats.json && webpack-bundle-analyzer stats.json",
+    "bundle-analyze": "run-s clean fetch content && webpack --config webpack.prod.js --node-env production && run-s printable content && webpack --config webpack.ssg.js --node-env production --env ssg --profile --json > stats.json && webpack-bundle-analyzer stats.json",
+    "fetch-repos": "node src/utilities/fetch-package-repos.js",
     "fetch": "run-p fetch:*",
     "fetch:readmes": "node src/utilities/fetch-package-readmes.js",
     "fetch:supporters": "node src/utilities/fetch-supporters.js",
-    "fetch:starter-kits": "node src/utilities/fetch-starter-kits.js",
     "prebuild": "npm run clean",
-    "build": "run-s fetch printable content && cross-env NODE_ENV=production webpack --config webpack.ssg.js && run-s clean-printable content && cross-env NODE_ENV=production webpack --config webpack.prod.js",
+    "build": "run-s fetch-repos fetch content && webpack --config webpack.prod.js --node-env production && run-s printable content && webpack --config webpack.ssg.js --node-env production --env ssg",
     "postbuild": "npm run sitemap",
-    "build-test": "npm run build && http-server dist/",
+    "build-test": "npm run build && http-server --port 4200 dist/",
     "test": "npm run lint",
     "lint": "run-s lint:*",
-    "lint:js": "eslint src --ext .js,.jsx,.md --cache true --cache-location .cache/.eslintcache",
-    "lint:markdown": "markdownlint --rules markdownlint-rule-emphasis-style --config ./.markdownlint.json *.md ./src/content/**/*.md --ignore './src/content/**/_*.md'",
+    "lint:js": "npm run lint-js .",
+    "lint-js": "eslint --cache --cache-location .cache/.eslintcache",
+    "lint:markdown": "npm run lint-markdown *.md ./src/content/**/*.md",
+    "lint-markdown": "markdownlint --config ./.markdownlint.json --ignore './src/content/**/_*.md' --ignore '.vale/**/*.md' --ignore '.github/**/*.md'",
     "lint:social": "alex . -q",
-    "lint:prose": "cp .proselintrc ~/ && proselint src/content",
-    "lint:links": "hyperlink -c 8 -r dist/index.html --canonicalroot https://webpack.js.org/ -i --skip https://img.shields.io --skip **/printable** --skip https://david-dm.org --skip https://codecov.io/gh --skip 'content-type-mismatch https://travis-ci.org' > internal-links.tap; cat internal-links.tap | tap-spot",
+    "lint:prose": "vale --config='.vale.ini' src/content",
+    "lint:links": "hyperlink -c 8 --root dist -r dist/index.html --canonicalroot https://webpack.js.org/ --internal --skip /plugins/extract-text-webpack-plugin/ --skip /printable --skip https:// --skip http:// --skip sw.js > internal-links.tap; cat internal-links.tap | tap-spot",
     "sitemap": "cd dist && sitemap-static --ignore-file=../sitemap-ignore.json --pretty --prefix=https://webpack.js.org/ > sitemap.xml",
     "serve": "npm run build && sirv start ./dist --port 4000",
-    "deploy": "gh-pages -d dist",
     "preprintable": "npm run clean-printable",
     "printable": "node ./src/scripts/concatenate-docs.js",
-    "jest": "jest"
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "lint-staged"
-    }
+    "jest": "jest",
+    "cypress:open": "cypress open",
+    "cypress:run": "cypress run",
+    "cypress:ci": "start-server-and-test build-test http://localhost:4200 cypress:run",
+    "prettier": "prettier --write '**/*.{js,json,jsx,css,scss,md,mdx}'",
+    "prepare": "husky install"
   },
   "lint-staged": {
     "*.{js,jsx,md}": [
-      "npm run lint:js"
+      "npm run lint-js"
     ],
     "*.md": [
-      "npm run lint:markdown"
+      "npm run lint-markdown"
+    ],
+    "*.{js,jsx,css,scss,md,mdx,json}": [
+      "prettier --write"
     ]
   },
-  "bundlesize": [
-    {
-      "path": "./dist/index.bundle.js",
-      "maxSize": "200 kB"
-    }
-  ],
   "devDependencies": {
-    "@mdx-js/loader": "0.15.7",
-    "@mdx-js/mdx": "0.15.7",
-    "@octokit/rest": "^16.27.1",
+    "@babel/core": "^7.13.8",
+    "@babel/eslint-parser": "^7.13.10",
+    "@babel/plugin-proposal-class-properties": "^7.12.1",
+    "@babel/preset-env": "^7.13.12",
+    "@babel/preset-react": "^7.12.13",
+    "@mdx-js/loader": "^1.6.21",
+    "@octokit/auth-action": "^1.3.2",
+    "@octokit/rest": "^18.5.0",
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.4.2",
+    "@svgr/webpack": "^5.5.0",
     "alex": "^5.1.0",
-    "autoprefixer": "^7.2.3",
-    "babel-core": "^6.26.0",
-    "babel-eslint": "^7.2.3",
-    "babel-loader": "^7.1.2",
-    "babel-plugin-syntax-dynamic-import": "^6.18.0",
-    "babel-plugin-transform-class-properties": "^6.24.1",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-preset-env": "^1.6.0",
-    "babel-preset-react": "^6.24.1",
-    "bundlesize": "^0.17.1",
-    "copy-webpack-plugin": "4.5.2",
-    "cross-env": "5.2.0",
-    "css-loader": "^0.28.10",
-    "directory-tree": "2.2.0",
-    "directory-tree-webpack-plugin": "0.3",
+    "autoprefixer": "^10.2.5",
+    "babel-loader": "^8.2.2",
+    "copy-webpack-plugin": "^8.1.0",
+    "css-loader": "^5.2.0",
+    "css-minimizer-webpack-plugin": "^1.3.0",
+    "cypress": "^6.8.0",
+    "directory-tree": "^2.2.7",
+    "directory-tree-webpack-plugin": "^1.0.2",
     "duplexer": "^0.1.1",
-    "eslint": "4.19.1",
-    "eslint-loader": "^2.0.0",
-    "eslint-plugin-markdown": "1.0.0-rc.0",
-    "file-loader": "^1.1.11",
-    "fontgen-loader": "git://github.com/EugeneHlushko/fontgen-loader.git#0.1.3",
-    "front-matter": "^2.3.0",
-    "gh-pages": "^1.0.0",
-    "html-webpack-plugin": "^3.2.0",
-    "html-webpack-template": "^6.1.0",
-    "http-server": "^0.10.0",
-    "husky": "^1.0.0-rc.8",
-    "hyperlink": "^4.4.0",
-    "jest": "^23.6.0",
-    "lint-staged": "^8.1.0",
-    "loader-utils": "^1.1.0",
-    "lodash": "^4.17.4",
-    "markdown-loader": "^4.0.0",
-    "markdownlint": "^0.11.0",
-    "markdownlint-cli": "^0.13.0",
-    "markdownlint-rule-emphasis-style": "^1.0.0",
-    "marked": "0.6.1",
-    "mini-css-extract-plugin": "^0.5.0",
-    "minimist": "1.2.3",
-    "mkdirp": "^0.5.1",
+    "eslint": "^7.22.0",
+    "eslint-config-prettier": "^8.1.0",
+    "eslint-plugin-cypress": "^2.11.2",
+    "eslint-plugin-markdown": "^2.0.0",
+    "eslint-plugin-mdx": "^1.11.1",
+    "eslint-plugin-react": "^7.23.1",
+    "eslint-plugin-react-hooks": "^4.2.0",
+    "front-matter": "^4.0.2",
+    "html-loader": "^2.1.2",
+    "html-webpack-plugin": "^5.3.0",
+    "http-server": "^0.12.3",
+    "husky": "^5.2.0",
+    "hyperlink": "^4.6.0",
+    "jest": "^26.6.3",
+    "lint-staged": "^10.5.4",
+    "lodash": "^4.17.21",
+    "markdownlint": "^0.23.1",
+    "markdownlint-cli": "^0.27.1",
+    "mini-css-extract-plugin": "^1.3.9",
+    "mkdirp": "^1.0.4",
     "modularscale-sass": "^3.0.3",
-    "node-sass": "^4.5.3",
+    "node-fetch": "^2.6.1",
     "npm-run-all": "^4.1.1",
-    "offline-plugin": "^5.0.7",
-    "optimize-css-assets-webpack-plugin": "^5.0.1",
-    "postcss-loader": "^2.1.3",
-    "redirect-webpack-plugin": "^0.1.1",
-    "remark": "^10.0.1",
-    "remark-autolink-headings": "5.2.1",
-    "remark-custom-blockquotes": "1.0.0",
+    "postcss": "^8.2.8",
+    "postcss-loader": "^5.2.0",
+    "prettier": "^2.2.1",
+    "react-refresh": "^0.9.0",
+    "redirect-webpack-plugin": "^1.0.0",
+    "remark": "^13.0.0",
+    "remark-autolink-headings": "^6.0.1",
+    "remark-emoji": "^2.1.0",
     "remark-extract-anchors": "1.1.1",
-    "remark-loader": "^0.3.0",
-    "remark-refractor": "1.1.0",
-    "remark-responsive-tables": "1.0.0",
-    "remark-slug": "^5.0.0",
-    "request-promise": "4.2.4",
-    "rimraf": "^2.6.2",
-    "sass-loader": "^6.0.6",
-    "sirv-cli": "^0.1.2",
+    "remark-frontmatter": "^3.0.0",
+    "remark-gfm": "^1.0.0",
+    "remark-html": "^13.0.1",
+    "remark-loader": "^3.0.0",
+    "remark-refractor": "git://github.com/montogeek/remark-refractor.git",
+    "remark-slug": "^6.0.0",
+    "rimraf": "^3.0.2",
+    "sass": "^1.32.8",
+    "sass-loader": "^11.0.1",
+    "sirv-cli": "^1.0.11",
     "sitemap-static": "^0.4.2",
+    "start-server-and-test": "^1.12.1",
     "static-site-generator-webpack-plugin": "^3.4.1",
-    "style-loader": "^0.18.2",
+    "style-loader": "^2.0.0",
     "tap-spot": "^1.1.1",
-    "terser-webpack-plugin": "^2.3.5",
-    "through2": "^2.0.3",
-    "uglifyjs-webpack-plugin": "^1.1.6",
-    "webpack": "^4.42.0",
-    "webpack-bundle-analyzer": "^3.3.2",
-    "webpack-cli": "^3.2.3",
-    "webpack-dev-server": "^3.2.1",
-    "webpack-merge": "^4.1.0"
+    "unist-util-visit": "^2.0.3",
+    "webpack": "^5.28.0",
+    "webpack-bundle-analyzer": "^4.4.0",
+    "webpack-cli": "^4.5.0",
+    "webpack-dev-server": "^4.0.0-beta.1",
+    "webpack-merge": "^5.7.3",
+    "workbox-webpack-plugin": "^6.1.2"
   },
   "dependencies": {
-    "ajv": "^5.5.2",
     "docsearch.js": "^2.5.2",
-    "gitter-sidecar": "^1.2.3",
-    "handlebars": "4.5.3",
-    "javascriptstuff-db": "^1.12.0",
-    "lodash.throttle": "^4.1.1",
+    "path-browserify": "^1.0.1",
     "prop-types": "^15.7.2",
-    "react": "^16.8.6",
+    "react": "^17.0.2",
     "react-banner": "^1.0.0-rc.0",
-    "react-document-title": "^2.0.3",
-    "react-dom": "^16.8.6",
+    "react-dom": "^17.0.2",
     "react-g-analytics": "0.4.2",
-    "react-hot-loader": "^4.0.0-beta.12",
-    "react-markdown": "4.2.2",
-    "react-router-dom": "^4.2.2",
-    "react-tiny-popover": "3.4.2",
+    "react-helmet-async": "^1.0.9",
+    "react-router-dom": "^5.2.0",
+    "react-spring": "next",
+    "react-tiny-popover": "5",
+    "react-use": "^17.2.1",
     "react-visibility-sensor": "^5.0.2",
-    "webpack-pwa-manifest": "^4.0.0",
+    "unist-util-visit": "^2.0.3",
+    "webpack-pwa-manifest": "^4.3.0",
     "webpack.vote": "https://github.com/webpack/voting-app.git",
-    "whatwg-fetch": "^2.0.3"
+    "workbox-window": "^6.1.2"
+  },
+  "resolutions": {
+    "sitemap-static/minimist": "1.2.5",
+    "ini": "1.3.7",
+    "eval": "^0.1.5"
   }
 }
