@@ -1,20 +1,71 @@
-import React from 'react';
-import ReactMarkdown from 'react-markdown';
+import { Children, isValidElement } from 'react';
 import { Details } from './components';
 
-const detailComponentsList = ['link', 'mode', 'entry', 'filename', 'publicPath', 'advancedOutput', 'expert', 'advancedModule', 'alias', 'advancedResolve', 'hints', 'devtool', 'target', 'externals', 'stats', 'advanced', 'libraryTarget'];
-
-const Pre = props => {
-  const newChildren = React.Children.map(props.children.props.children, child => {
-    if (React.isValidElement(child)) {
-      if (child.props.props.className.includes('keyword')) {
-        if (!detailComponentsList.includes(child.props.props.componentname)) return child;
-
-        return <Details children={child.props.children.slice(4, React.Children.count(child.props.children) - 4)} url={child.props.props.url} />;
-      }
-
-      if (child.props.props.className.includes('comment')) {
-        return <ReactMarkdown source={child.props.children} disallowedTypes={['paragraph']} unwrapDisallowed />;
+const detailComponentsList = [
+  'link',
+  'mode',
+  'entry',
+  'path',
+  'filename',
+  'publicPath',
+  'library',
+  'libraryType',
+  'libraryName',
+  'advancedLibrary',
+  'advancedOutput',
+  'expertOutput',
+  'expertOutputB',
+  'expert',
+  'advancedConditions',
+  'moduleType',
+  'advancedActions',
+  'advancedModule',
+  'modules',
+  'alias',
+  'advancedResolve',
+  'expertResolve',
+  'hints',
+  'devtool',
+  'target',
+  'externals',
+  'externalsType',
+  'externalsPresets',
+  'ignoreWarnings',
+  'stats',
+  'preset',
+  'advancedGlobal',
+  'advancedAssets',
+  'advancedChunkGroups',
+  'advancedChunks',
+  'advancedModules',
+  'expertModules',
+  'advancedStatsOptimization',
+  'advancedOptimization',
+  'cacheGroupAdvancedSelectors',
+  'cacheGroupAdvancedEffects',
+  'advancedSelectors',
+  'advancedEffects',
+  'fallbackCacheGroup',
+  'advanced',
+  'advancedCaching',
+  'advancedBuild',
+];
+export const Pre = (props) => {
+  // eslint-disable-next-line
+  const newChildren = Children.map(props.children.props.children, (child) => {
+    if (isValidElement(child)) {
+      if (child.props.className.includes('keyword')) {
+        if (!detailComponentsList.includes(child.props.componentname))
+          return child;
+        return (
+          <Details
+            myChilds={child.props.children.slice(
+              4,
+              Children.count(child.props.children) - 4
+            )}
+            url={child.props.url}
+          />
+        );
       }
     }
 
@@ -22,7 +73,7 @@ const Pre = props => {
   });
 
   const newProps = {
-    children: newChildren
+    children: newChildren,
   };
 
   return (
@@ -31,9 +82,4 @@ const Pre = props => {
     </pre>
   );
 };
-
-export default {
-  components: {
-    pre: Pre
-  }
-};
+Pre.propTypes = {};
