@@ -1,10 +1,8 @@
-import React from 'react';
 import './Markdown.scss';
+import PropTypes from 'prop-types';
 
-const Markdown = props => (
-  <div className="markdown">
-    { props.children }
-  </div>
-);
-
+const Markdown = (props) => <div className="markdown">{props.children}</div>;
+Markdown.propTypes = {
+  children: PropTypes.node,
+};
 export default Markdown;
