@@ -1,5 +1,4 @@
 // Import External Dependencies
-import React from 'react';
 import Link from '../Link/Link';
 
 // Imports for Ad Content
@@ -10,24 +9,20 @@ import WebpackIcon from '../../assets/icon-square-small.svg';
 import '../Sponsors/Sponsors.scss';
 
 // Create and export the component
-export default props => {
+const Sponsors = () => {
   return (
     <div className="sponsors">
       <Link
         className="sponsors__content"
-        to="https://www.ag-grid.com/?utm_source=webpack&utm_medium=banner&utm_campaign=sponsorship">
-        <img src={ AGLogo } />
-        <img src={ WebpackIcon } style={{ width: '100px' }} />
-        <div style={{
-          marginTop: '1em',
-          fontSize: '2em',
-          textAlign: 'center',
-          color: '#535353'
-        }}>
+        to="https://www.ag-grid.com/?utm_source=webpack&utm_medium=banner&utm_campaign=sponsorship"
+      >
+        <img src={AGLogo} />
+        <img className="sponsors__img__webpack" src={WebpackIcon} />
+        <div className="sponsors__content-cont">
           ag-grid is proud to partner with webpack
         </div>
-
       </Link>
     </div>
   );
 };
+export default Sponsors;
