@@ -7,35 +7,46 @@ export default [
     avatar: 'https://static.moonmail.io/moonmail-logo.svg',
     website: 'https://moonmail.io/?utm_source=webpack.js.org',
     totalDonations: 11000,
-    reason: 'Paypal'
+    monthlyDonations: 0,
+    reason: 'Paypal',
   },
   {
     name: 'Google Angular',
-    avatar: 'https://res.cloudinary.com/opencollective/image/upload/v1485288529/angular_uxllte.png',
-    website: 'https://angular.io/?utm_source=webpack&utm_medium=documentation&utm_campaign=sponsorship',
+    avatar:
+      'https://res.cloudinary.com/opencollective/image/upload/v1485288529/angular_uxllte.png',
+    website:
+      'https://angular.io/?utm_source=webpack&utm_medium=documentation&utm_campaign=sponsorship',
     totalDonations: 250000,
-    reason: 'Paypal'
+    monthlyDonations: 0,
+    reason: 'Paypal',
   },
   {
     name: 'Architects.io',
     avatar: null,
-    website: 'http://architects.io/?utm_source=webpack&utm_medium=documentation&utm_campaign=sponsorship',
+    website:
+      'http://architects.io/?utm_source=webpack&utm_medium=documentation&utm_campaign=sponsorship',
     totalDonations: 30000,
-    reason: 'Paypal'
+    monthlyDonations: 0,
+    reason: 'Paypal',
   },
   {
     slug: 'peerigon',
     name: 'Peerigon',
-    avatar: 'https://opencollective-production.s3-us-west-1.amazonaws.com/e8a1de10-99c8-11e6-8650-f92e594d5de8.png',
-    website: 'https://peerigon.com/?utm_source=webpack&utm_medium=documentation&utm_campaign=sponsorship',
+    avatar:
+      'https://opencollective-production.s3-us-west-1.amazonaws.com/e8a1de10-99c8-11e6-8650-f92e594d5de8.png',
+    website:
+      'https://peerigon.com/?utm_source=webpack&utm_medium=documentation&utm_campaign=sponsorship',
     totalDonations: 144139,
-    reason: 'webpack meetup 2017-07'
+    monthlyDonations: 0,
+    reason: 'webpack meetup 2017-07',
   },
   {
     name: 'Segment',
     avatar: SegmentLogo,
-    website: 'https://segment.com/?utm_source=webpack&utm_medium=documentation&utm_campaign=sponsorship',
+    website:
+      'https://segment.com/?utm_source=webpack&utm_medium=documentation&utm_campaign=sponsorship',
     totalDonations: 2400000,
-    reason: 'Sponsorship 2017-07 - 2017-09'
-  }
+    monthlyDonations: 0,
+    reason: 'Sponsorship 2017-07 - 2017-09',
+  },
 ];
