@@ -1,14 +1,14 @@
 // Import External Dependencies
-import React from 'react';
+import { Children, cloneElement, PureComponent } from 'react';
 import PropTypes from 'prop-types';
 
 // Load Styling
 import './TextRotater.scss';
 
-export default class TextRotater extends React.PureComponent {
+export default class TextRotater extends PureComponent {
   static defaultProps = {
     delay: 0,
-    repeatDelay: 3000
+    repeatDelay: 3000,
   };
 
   static propTypes = {
@@ -17,40 +17,37 @@ export default class TextRotater extends React.PureComponent {
     repeatDelay: PropTypes.number,
     // Needed to prevent jump when
     // rotating between texts of different widths
-    maxWidth: PropTypes.number
+    maxWidth: PropTypes.number,
   };
 
   state = {
     currentIndex: 0,
-    contentHeight: 0
-  }
+    contentHeight: 0,
+  };
 
   render() {
     const { children, maxWidth } = this.props;
     const { currentIndex, contentHeight } = this.state;
-    const childrenCount = React.Children.count(children);
+    const childrenCount = Children.count(children);
 
-    const currentChild = React.cloneElement(
-      children[currentIndex], {
-        ref: c => (this.content = c)
-      }
-    );
+    const currentChild = cloneElement(children[currentIndex], {
+      ref: (c) => (this.content = c),
+    });
 
-    const nextChild = React.cloneElement(
-      children[
-        (currentIndex + 1) % childrenCount
-      ]
+    const nextChild = cloneElement(
+      children[(currentIndex + 1) % childrenCount]
     );
 
     return (
       <div className="text-rotater">
         <div
           className="text-rotater__wrap"
-          ref={ trw => (this.textRotatorWrap = trw) }
-          onTransitionEnd={ this._handleTransitionEnd }
-          style={ { height: contentHeight, width: maxWidth } }>
-          { currentChild }
-          { nextChild }
+          ref={(trw) => (this.textRotatorWrap = trw)}
+          onTransitionEnd={this._handleTransitionEnd}
+          style={{ height: contentHeight, width: maxWidth }}
+        >
+          {currentChild}
+          {nextChild}
         </div>
       </div>
     );
@@ -69,24 +66,18 @@ export default class TextRotater extends React.PureComponent {
       }
     }, delay);
 
-    window.addEventListener(
-      'resize',
-      this._calculateContentHeight
-    );
+    window.addEventListener('resize', this._calculateContentHeight);
   }
 
   componentWillUnmount() {
-    window.removeEventListener(
-      'resize',
-      this._calculateContentHeight
-    );
+    window.removeEventListener('resize', this._calculateContentHeight);
   }
 
   _calculateContentHeight = () => {
     this.setState({
       contentHeight: this.content.clientHeight,
     });
-  }
+  };
 
   _handleTransitionEnd = () => {
     const { children, repeatDelay } = this.props;
@@ -94,15 +85,19 @@ export default class TextRotater extends React.PureComponent {
     if (this.textRotatorWrap) {
       this.textRotatorWrap.classList.remove('text-rotater--slide-up');
 
-      this.setState({
-        currentIndex: (this.state.currentIndex + 1) % React.Children.count(children),
-      }, () => {
-        setTimeout(() => {
-          if (this.textRotatorWrap) {
-            this.textRotatorWrap.classList.add('text-rotater--slide-up');
-          }
-        }, repeatDelay);
-      });
+      this.setState(
+        {
+          currentIndex:
+            (this.state.currentIndex + 1) % Children.count(children),
+        },
+        () => {
+          setTimeout(() => {
+            if (this.textRotatorWrap) {
+              this.textRotatorWrap.classList.add('text-rotater--slide-up');
+            }
+          }, repeatDelay);
+        }
+      );
     }
-  }
+  };
 }
