@@ -1,17 +1,17 @@
-import React from 'react';
+import { Component } from 'react';
 import Container from '../Container/Container';
 import 'webpack.vote/dist/style.min.css';
 
 // Load Styling
 import './Vote.scss';
 
-export default class Vote extends React.Component {
+export default class Vote extends Component {
   state = {
-    VoteApp: null
-  }
+    VoteApp: null,
+  };
   componentDidMount() {
     if (window && !window.__ssgrun) {
-      import('webpack.vote').then(VoteApp => {
+      import('webpack.vote').then((VoteApp) => {
         this.setState({ VoteApp: VoteApp.default || VoteApp });
       });
     }
@@ -20,7 +20,7 @@ export default class Vote extends React.Component {
     const { VoteApp } = this.state;
     return (
       <Container className="vote markdown">
-        { VoteApp ? <VoteApp development={ false } /> : null }
+        {VoteApp ? <VoteApp development={false} /> : null}
       </Container>
     );
   }
