@@ -1,14 +1,13 @@
 // Import External Dependencies
-import React from 'react';
 import ReactDOM from 'react-dom';
-import { BrowserRouter, Route } from 'react-router-dom';
+import { BrowserRouter } from 'react-router-dom';
 import { BrowserRouter as AnalyticsRouter } from 'react-g-analytics';
 
-// Import Components
-import Site from './components/Site/Site';
+import App from './App.jsx';
 
 // Import helpers
 import isClient from './utilities/is-client';
+import { HelmetProvider } from 'react-helmet-async';
 
 const isProduction = process.env.NODE_ENV === 'production';
 
@@ -17,15 +16,12 @@ const render = isProduction ? ReactDOM.hydrate : ReactDOM.render;
 
 // Client Side Rendering
 if (isClient) {
-  render((
+  render(
     <Router id="UA-46921629-2">
-      <Route
-        path="/"
-        render={ props => (
-          <Site
-            { ...props }
-            import={ path => import(`./content/${path}`) } />
-        )} />
-    </Router>
-  ), document.getElementById('root'));
+      <HelmetProvider>
+        <App />
+      </HelmetProvider>
+    </Router>,
+    document.getElementById('root')
+  );
 }
