@@ -4,9 +4,13 @@ const frontMatter = require('front-matter');
 const remark = require('remark');
 const slug = require('remark-slug');
 const extractAnchors = require('remark-extract-anchors');
+const remarkHtml = require('remark-html');
+// why we have remark-frontmatter now
+// see https://github.com/webpack/webpack.js.org/pull/4111/files#r517309746
+const frontmatter = require('remark-frontmatter');
+const gfm = require('remark-gfm');
 
 const enhance = (tree, options) => {
-
   // delete `./` root directory on node
   const dir = path.normalize(options.dir).replace(/^(\.\/)/gm, '');
 
@@ -39,8 +43,12 @@ const enhance = (tree, options) => {
 
     remark()
       .use(slug)
+      .use(frontmatter)
+      .use(gfm)
+      .use(require('remark-emoji'))
       .use(extractAnchors, { anchors, levels: 3 })
-      .process(content, err => {
+      .use(remarkHtml)
+      .process(content, (err) => {
         if (err) {
           throw err;
         }
@@ -48,13 +56,17 @@ const enhance = (tree, options) => {
 
     tree.anchors = anchors;
 
-    Object.assign(tree, {
-      path: tree.path.replace(/\\/g, '/')
-    }, attributes);
+    Object.assign(
+      tree,
+      {
+        path: tree.path.replace(/\\/g, '/'),
+      },
+      attributes
+    );
   }
 };
 
-const filter = item => true;
+const filter = () => true;
 
 const sort = (a, b) => {
   let group1 = (a.group || '').toLowerCase();
@@ -76,7 +88,7 @@ function restructure(item, options) {
   enhance(item, options);
 
   if (item.children) {
-    item.children.forEach(child => restructure(child, options));
+    item.children.forEach((child) => restructure(child, options));
 
     item.children.filter(filter);
     item.children.sort(sort);
@@ -89,5 +101,5 @@ module.exports = {
   enhance,
   filter,
   restructure,
-  sort
+  sort,
 };
