@@ -2,8 +2,8 @@ const _ = require('lodash');
 const fs = require('fs');
 const path = require('path');
 const { promisify } = require('util');
-const mkdirp = promisify(require('mkdirp'));
-const request = require('request-promise');
+const mkdirp = require('mkdirp');
+const fetch = require('node-fetch');
 
 const yamlHeadmatter = require('./yaml-headmatter.js');
 const processReadme = require('./process-readme.js');
@@ -16,7 +16,7 @@ const types = ['loaders', 'plugins'];
 
 const pathMap = {
   loaders: path.resolve(__dirname, '../content/loaders'),
-  plugins: path.resolve(__dirname, '../content/plugins')
+  plugins: path.resolve(__dirname, '../content/plugins'),
 };
 
 async function main() {
@@ -25,10 +25,12 @@ async function main() {
 
     await mkdirp(outputDir);
 
-    const repos = JSON.parse(await readFile(path.resolve(__dirname, `../../repositories/${type}.json`)));
+    const repos = JSON.parse(
+      await readFile(path.resolve(__dirname, `../../repositories/${type}.json`))
+    );
 
     for (const repo of repos) {
-      const [org, packageName] = repo.split('/');
+      const [, packageName] = repo.split('/');
       const url = `https://raw.githubusercontent.com/${repo}/master/README.md`;
       const htmlUrl = `https://github.com/${repo}`;
       const editUrl = `${htmlUrl}/edit/master/README.md`;
@@ -47,23 +49,16 @@ async function main() {
         title: title,
         source: url,
         edit: editUrl,
-        repo: htmlUrl
+        repo: htmlUrl,
       });
 
-      request(url)
-        .then(async content => {
-          const body = processReadme(content, { source: url });
-
-          await writeFile(fileName, headmatter + body);
-
-          console.log('Generated:', path.relative(cwd, fileName));
-        })
-        .catch(err => {
-          throw err;
-        });
+      const response = await fetch(url);
+      const content = await response.text();
+      const body = processReadme(content, { source: url });
+      await writeFile(fileName, headmatter + body);
+      console.log('Generated:', path.relative(cwd, fileName));
     }
   }
 }
 
 main();
-
